Callback for identifiers.
This change makes it easier to process method parameters as well as removing
the need of having to use the array based Sexp output for for example variable
assignment.

In this commit I've also changed all getter methods in the token classes to
accessors. This makes it possible to modify certain data of a token without
having to create an entirely new instance.

Signed-off-by: Yorick Peterse <yorickpeterse@gmail.com>