Started rewriting the parser from scratch.
While developing the old parser (moved to parser_old.rb for the time being) I
realized that the code could be simplified to such an extent that it becomes
less of a pain to work with.

I also decided to split up the parser tests into different files instead of one
big file, this in turn should make it easier to write and maintain tests.

Signed-off-by: Yorick Peterse <yorickpeterse@gmail.com>