Initial steps for parsing attribute assignments.
Currently the parser has limited capabilities of parsing object attribute
assignments and references.

While working on this feature I realized that the use of a VariableToken
instance for assignments doesn't make a whole lot of sense since you're not
always assigning data to a variable (as is the case with attributes). In the
following commit(s) I'll introduce a new token class, probably called
AssignmentToken, that will replace the VariableToken class in these cases.

Signed-off-by: Yorick Peterse <yorickpeterse@gmail.com>