Change for iterator class structures.
The classes used for iterator events no longer require an #initialize() method
that accepts a file name. Instead this file name is passed to individual
callback methods as a second parameter.

This means that instead of this:

    class Callback
      def initialize(file)
        @file = file
      end

      def on_class(token)

      end
    end

You'd write this:

    class Callback
      def on_class(token, file)

      end
    end

This change also means the class Rlint::Iterator itself has become easier to
re-use for different files without having to create a new instance for each
file.

Signed-off-by: Yorick Peterse <yorickpeterse@gmail.com>