Method call validation.
General
-------

Rlint is now capable of validating method calls using
Rlint::Analyze::Definitions. For example, it knows the following code is
invalid:

    name = 'Ruby'

    name.downcasex

It also knows the following is perfectly valid:

    'ruby'.downcase

To make things even better it also knows the difference between a method call
using an instance and class method (though it's not entirely fool proof).

However, there are still quite a few things that still have to be taken care
of such as proper class inheritence (and testing this), dealing with
method_missing, method arity validation and so on.

Rlint::Symbols
--------------

The constant Rling::SYMBOLS has been removed along with the code that generated
it. Instead of pre-generating this data it is generated when Rlint is loaded.
The benefit of this is that it's much easier to maintain and removes the need
of having to keep a 15k LOC file around. The drawback is that Rlint now uses
around 16 MB of base RAM instead of around 7MB.

Signed-off-by: Yorick Peterse <yorickpeterse@gmail.com>