Multiple parent scopes and parameter scoping.
Scope instances can now have multiple parent scopes. This is especially useful
for classes that should inherit data from the parent class as well as the
global scope.

Rlint::Analyze::Definitions now also takes care of method parameters and adds
these as local variables in the scope for the method.

Signed-off-by: Yorick Peterse <yorickpeterse@gmail.com>