Inherit constants inside namespaces.
When parsing code that is defined inside a namespace the various constants of
the parent namespace segments should be inherited. Take for example the
following code:

    module A
      module B
        def self.example

        end
      end
    end

    module A
      module C
        B.example
      end
    end

Here `A::B` is called inside `A::C` but without the full namespace path being
specified. Rlint is now capable of dealing with code such as this.

Signed-off-by: Yorick Peterse <yorickpeterse@gmail.com>