Rewrote the parsing of mass/splat assignments.
When parsing mass assignments RubyLint::Parser now returns an array of
AssignmentToken instances instead of a single instance containing multiple sub
instances.

Another change is that the parser also handles the task of properly assigning
the values to the corresponding variables when mass assigning variables. This
also works when splat assigning such as the following example:

    number, *this_is_an_empty_array = 10

This would result in an AST similar to the following:

    [
        #<RubyLint::Token::AssignmentToken
        @name="number"
        @value=#<RubyLint::Token::Token @type=:integer @value="10"> ...>,

        #<RubyLint::Token::AssignmentToken
        @name="this_is_an_empty_array"
        @value=nil ...>,
    ]

Signed-off-by: Yorick Peterse <yorickpeterse@gmail.com>