Properly resolve method receiver definitions.
The various parts of ruby-lint that deal with method calls on receiver
objects/constants are now able to do so without breaking things. Having said
that, I'm currently not very happy with the code of
RubyLint::Helper::Scoping#resolve_method_receiver as I feel it's too complex
and tries to do too much on its own. I plan on breaking this method into
smaller methods and cleaning it up in the following commits.

Signed-off-by: Yorick Peterse <yorickpeterse@gmail.com>