Started completely re-writing the AST.
The current AST that's being generated by RubyLint::Parser is overly complex
and confusing due to the large number of classes used for different node types.
After having a discussion about ASTs and the likes with @whitequark I decided
that the AST generated by the parser has to be re-written from scratch.

To make things easier I'm using "Furnace" which provides a simple, immutable
class that can be used for representing nodes in an AST. Along with changing
the AST will come various changes to the way the definitions list is built as
well as how callback classes work (due to different event names being used).
None of this will be backwards compatible with what I've currently pushed to
Rubygems but that's expected when something is still alpha quality software.

Signed-off-by: Yorick Peterse <yorickpeterse@gmail.com>