First steps for a new arguments AST.
The current arguments AST is not something I'm too happy with and as such I'm
reworking it. In this particular commit the new AST looks like the following:

    s(:arguments, s(:required_arguments, s(:local_variable, 'foo')))

However, since this AST is both used for method calls and definitions as well
as for array/field references it becomes a bit painful to work with. Most
likely I will implement the following AST in the coming days:

    s(
        :arguments,
        s(:required_argument, 'foo'),
        s(:optional_argument, 'bar', s(:integer, '10'))
    )

Since method parameters can only be local variables (or "identifiers" as Ripper
calls them in this case) there's no particular benefit to storing them as sub
nodes. The above AST would greatly simplify things without needing an overly
verbose AST.

Signed-off-by: Yorick Peterse <yorickpeterse@gmail.com>