Removed RubyLint::Importer.
The Importer module has been removed in favour of manually defining the various
globally available constants and methods. Although this approach is much more
time consuming from a developers perspective it does offer a much more flexible
way of doing it as well as being more reliable and faster in terms of execution
time.

Signed-off-by: Yorick Peterse <yorickpeterse@gmail.com>