Release 0.0.3.
Version 0.0.3 is a *massive* difference compared to the previous release.
Almost everything has been re-written from the ground up, lots of bugs have
been fixed (and new ones were introduced), analysis is better, the CLI is now
actually usable and documentation is much better.

However, I've kept the version number so small for a good reason: false
positives. Although ruby-lint is getting more and more mature with every commit
there's still *a lot* of work to be done.

For the new release I'll be focusing on replacing the in house parser with
<https://github.com/whitequark/parser>. Using this parser would introduce
support for Ruby 2.0 as well as making it able to get rid of a lot of code and
the potentially corresponding problems. In other words, I can actually focus on
analysis instead of parsing *and* analysis.

For the upcoming release I also plan on breaking the DefinitionsBuilder class
up into smaller components. Although it's not terrible by any means it's a
complex class (for a complex process) with a lot of room for improvements.

Signed-off-by: Yorick Peterse <yorickpeterse@gmail.com>