Initial work for a ruby-lint "virtual machine".
This VM (which isn't entirely a VM but I like the name) will be the replacement
of the rather clunky RubyLint::DefinitionsBuilder class. The new process of
building definitions is broken up into smaller parts for specific cases (e.g.
classes and modules) and uses stacks for assigning/resolving variable values.

Signed-off-by: Yorick Peterse <yorickpeterse@gmail.com>