Migrated to RSpec.
Although Bacon has served me well for the time being the output of Bacon and
the way to extend it started to annoy me too much. Although RSpec is quite
heavy in terms of the testing suites it should do everything I'll ever need
without having to write weird hacks.