Extracting method information from YARD tags.
The changes in this commit allow ruby-lint to extract method information from
YARD tags found in the corresponding comments of the method. Currently only
`@param` and `@return` tags are supported for methods but more might be added
in the future (given they're useful for ruby-lint's internals).

Consider the following code:

    def double(number)
      return number * 2
    end

If ruby-lint were to analyze this code as is it would not know what type
`number` holds since there's no type hinting in Ruby. Now lets document the
method using YARD tags:

    # @param [Numeric] number
    def double(number)
      return number * 2
    end

Using this simple comment ruby-lint now knows that `number` is a `Numeric`
value. The same is the case for return values:

    # @param [Numeric] number
    # @return [Numeric]
    def double(number)
      return number * 2
    end

Currently only YARD is supported and its only used for methods. Compound values
such as `Array<String>` also work but the inner value (`String` in this case)
is ignored since ruby-lint doesn't support compound values (at least not yet).

Support for Rdoc could in theory be added but since I don't use Rdoc myself I
have little interest in this. The Rdoc syntax is also a bit more freeform and
thus harder to parse compared to YARD's stricter syntax.