Excluding Object methods in the inspector.
Instead of relying on #methods(..., false) and #instance_methods(..., false)
the inspector now creates a diff between Object and the constant it inspects
while also taking into account methods that exist in Object but are redefined
in the given constant.

This commit in turn means that various Rails methods that were previously
excluded (e.g. ActionController::Base.layout) are now available.