Added RubyLint::FileScanner.
This class can be used for finding Ruby files that belong to a certian Ruby
constant (or at least try to). For example, given the constant "User" it will
try to look for "user.rb" in a set of directories using a glob() pattern.

This class will be one of the building blocks of being able to find externally
defined constants, one of the requirements for #29.