Create definitions for non existing values.
This includes missing return values, `nil` references and various others. This
was primarily introduced to ensure that the amount of arguments processed by
VirtualMachine#after_send is consistent with the amount it should process. It
also removes the need for extra checks to see whether a particular bit of code
is dealing with a definition or a `nil` value.