Properly look up methods when in with block scopes
When dealing with block scopes a method should be looked up in the previous
scope if it can't be found in the current one. This is a bit of a hacky
solution but it requires the least amount of code changes. Ideally I'd have the
format changed in parser from `(block (send))` to `(send (block))` but I doubt
this will happen any time soon.