Use Symbol instances for callback names.
Instead of creating an Array and 2 String instances for every pair of callback
names the names should be generated as Symbol instances in
RubyLint::Iterator#iterate itself.

Although the performance gain from this is fairly small it's one of the many
changes I'll hopefully be able to make to eventually speed things up enough
that you can actually notice it.