Lets not enforce the latest Racc Gem.
I totally forgot that Racc is a C extension and that jruby doesn't really like
C extensions. This means that Travis CI and friends (and those that just want
to use jruby) would have to explicitly enable C extension support for jruby.

Instead I'll go down the alternative route of just requiring users to manually
update Racc if they're on a non MRI system and wish to have better performance.