Changed the default FileScanner directories.
Instead of using `$PWD as the default the FileScanner class now uses the
following two directories:

* $PWD/app
* $PWD/lib

Since these are the most common two Ruby application directories this provides
a sane default that doesn't have the drawbacks that using `$PWD` has. For
example, the latter wouldn't work very well if ruby-lint were to be executed
from a user's home directory.

See #81 for more information.