Don't glob when no directories are specified.
Before this commit when no directories were specified for the FileScanner class
(or no existing ones) the code would end up globbing for /**/*.rb. That is, it
would search for all Ruby source files starting in the root directory of your
disk. Yikes.

This commit fixes this by simply not globbing at all if the list of directories
is empty.

See #83 and #84 for more information.