Removed the GlobalState module.
This module has been replaced by RubyLint::Definition::Registry and a
corresponding singleton method RubyLint.registry.

This new registry is used to register definitions as well as applying them to
an existing RubyLint::Definition::RubyObject instance. This new setup allows
definitions to be created in a local state instead of a global one. The benefit
of this approach is that it becomes a lot easier to deal with analysis of
monkeypatches for loaded constants without having to resort to nasty hacks.

The downside of this approach is that if you have a long running process, such
as a web application, it would have to re-create the same definitions for every
request even if they're never modified. This is a change I'm willing to
introduce for the sake of code simplicity. If creating definitions is not fast
enough then I'll consider that a bug and try to fix it.