Only invoke call for Proc return values.
When a method call is evaluated using RubyObject#call this method should only
invoke `call` on Proc instances, not on any arbitrary object that happens to
respond to `call`. This is due to RubyObject instances responding to this
method but not always doing something with it. This particular code was only
meant to deal with Proc instances anyway.

This fixes #91.