Associate `send` argument nodes.
When evaluating method calls the corresponding argument nodes are associated
with their respecitve definitions. This means that the `UselessEqualityChecks
class no longer needs to call `VirtualMachine#evaluate_node`. The latter would
in certain cases raise an error as it would try to modify a frozen Hash.

Thanks to Peter Koppstein for reporting the issue of the frozen Hash errors.