Removed file loading debug output.
I've decided that ruby-lint will *not* feature a dedicated debugging API. The
problem with such an approach is that it requires certain hooks to be called
inside methods. Such a process is not really different from putting debug code
in directly when it's only used internally.

A more elegant solution, one that doesn't require the target code to be
modified, is overwriting methods where needed. I might add this in the future
but for now I'll leave it out.

This fixes #107.