Rip out the caching system.
The caching of ASTs has been too problematic for too long now. I even reached a
point where I would always use --disable-cache. In most cases the cache system
wouldn't improve performance, in some cases it would even make things slower.

Instead of caching things I'll be taking a closer look in the future at
optimizing the runtime itself. For example, right now the process of analyzing
multiple files is rather in-efficient.

Due to changes of the CLI, which I consider a public interface, I'll be bumping
the version of ruby-lint to 2.0.

This fixes #103.